requirements.txt
flask==2.2.2
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.3
python-dotenv==0.20.0
SQLAlchemy-serializer==1.4.1
SQLAlchemy==1.4.40